Parameter Descriptions for Package

Default parameter descriptions which may be overridden in individual functions.

Arguments

...

Unused.

args

A named list of the argument values.

dist

A string specifying the distribution in the fitdists object or "top" to use the distribution with most weight or "multi" to treat the distributions as a single distribution.

nrow

A count of the number of rows in the generated data which must be between 5 and 1,000,

nsim

A count of the number of data sets to generate.

pars

A named list of the parameter values.

replace

A flag specifying whether to sample with replacement.

seed

An integer of the starting seed or NULL.

start_sim

A count of the number of the simulation to start from.

stream

A count of the stream number.

Details

A flag is a non-missing logical scalar.

A string is a non-missing character scalar.