Generate Data for Simulations

Generate Data for Simulations

Usage

ssd_generate_data(x, ...)

# S3 method for class 'data.frame'
ssd_generate_data(x, ..., replace = FALSE, nrow = 6L, nsims = 100L)

# S3 method for class 'fitdists'
ssd_generate_data(x, ..., dist = "top", nrow = 6L, nsims = 100L)

# S3 method for class 'tmbfit'
ssd_generate_data(x, ..., nrow = 6L, nsims = 100L)

# S3 method for class 'character'
ssd_generate_data(x, ..., pars = list(), nrow = 6L, nsims = 100L)

# S3 method for class '`function`'
ssd_generate_data(x, ..., args = list(), nrow = 6L, nsims = 100L)

Arguments

x

The object to use for generating the data.

...

Unused.

replace

A flag specifying whether to sample with replacement.

nrow

A integer vector of the number of rows in the generated data.

nsims

A count of the number of data sets to generate.

dist

A string specifying the distribution in the fitdists object or "top" to use the distribution with most weight or "multi" to treat the distributions as a single distribution.

pars

A named list of the parameter values.

args

A named list of the argument values.

Value

A tibble of generate datasets.

Methods (by class)

ssd_generate_data(data.frame): Generate data from data.frame

ssd_generate_data(fitdists): Generate data from fitdists object

ssd_generate_data(tmbfit): Generate data from tmbfit object

ssd_generate_data(character): Generate data using distribution name

ssd_generate_data(`function`): Generate data using function

Examples

ssd_generate_data(ssddata::ccme_boron, nrow = 5, nsims = 3)
#> # A tibble: 15 × 9
#>    Chemical Species                   Conc Group Units   row   sim  nrow replace
#>    <chr>    <chr>                    <dbl> <fct> <chr> <int> <int> <dbl> <lgl>  
#>  1 Boron    Rana pipiens              20.4 Amph… mg/L      1     1     5 FALSE  
#>  2 Boron    Chlorella vulgaris         5.2 Plant mg/L      2     1     5 FALSE  
#>  3 Boron    Paramecium caudatum       20   Inve… mg/L      3     1     5 FALSE  
#>  4 Boron    Carassius auratus         15.6 Fish  mg/L      4     1     5 FALSE  
#>  5 Boron    Pimephales promelas       18.3 Fish  mg/L      5     1     5 FALSE  
#>  6 Boron    Bufo americanus           50   Amph… mg/L      1     2     5 FALSE  
#>  7 Boron    Lemna minor               60   Plant mg/L      2     2     5 FALSE  
#>  8 Boron    Brachydanio rerio         10   Fish  mg/L      3     2     5 FALSE  
#>  9 Boron    Pimephales promelas       18.3 Fish  mg/L      4     2     5 FALSE  
#> 10 Boron    Ceriodaphnia dubia        13.4 Inve… mg/L      5     2     5 FALSE  
#> 11 Boron    Carassius auratus         15.6 Fish  mg/L      1     3     5 FALSE  
#> 12 Boron    Lemna minor               60   Plant mg/L      2     3     5 FALSE  
#> 13 Boron    Selenastrum capricornut…  12.3 Plant mg/L      3     3     5 FALSE  
#> 14 Boron    Bufo americanus           50   Amph… mg/L      4     3     5 FALSE  
#> 15 Boron    Rana pipiens              20.4 Amph… mg/L      5     3     5 FALSE  

fit <- ssdtools::ssd_fit_dists(ssddata::ccme_boron)
ssd_generate_data(fit, nrow = 5, nsims = 3)
#> Error in glance.fitdists(x, wt = TRUE): `...` must be unused.

fit <- ssdtools::ssd_fit_dists(ssddata::ccme_boron)
ssd_generate_data(fit[[1]], nrow = 5, nsims = 3)
#> # A tibble: 15 × 5
#>      Conc   row   sim  nrow args            
#>     <dbl> <int> <int> <dbl> <list>          
#>  1  0.897     1     1     5 <named list [3]>
#>  2 15.9       2     1     5 <named list [3]>
#>  3  5.68      3     1     5 <named list [3]>
#>  4  2.93      4     1     5 <named list [3]>
#>  5 28.1       5     1     5 <named list [3]>
#>  6 10.2       1     2     5 <named list [3]>
#>  7 17.7       2     2     5 <named list [3]>
#>  8  1.95      3     2     5 <named list [3]>
#>  9 35.0       4     2     5 <named list [3]>
#> 10 19.1       5     2     5 <named list [3]>
#> 11 20.0       1     3     5 <named list [3]>
#> 12 65.4       2     3     5 <named list [3]>
#> 13 18.2       3     3     5 <named list [3]>
#> 14 14.8       4     3     5 <named list [3]>
#> 15  0.801     5     3     5 <named list [3]>

ssd_generate_data("lnorm", nrow = 5, nsims = 3)
#> # A tibble: 15 × 5
#>      Conc   row   sim  nrow args            
#>     <dbl> <int> <int> <dbl> <list>          
#>  1  0.731     1     1     5 <named list [1]>
#>  2  2.91      2     1     5 <named list [1]>
#>  3  1.07      3     1     5 <named list [1]>
#>  4  0.528     4     1     5 <named list [1]>
#>  5  0.951     5     1     5 <named list [1]>
#>  6  0.778     1     2     5 <named list [1]>
#>  7  1.56      2     2     5 <named list [1]>
#>  8 15.7       3     2     5 <named list [1]>
#>  9  1.05      4     2     5 <named list [1]>
#> 10  1.78      5     2     5 <named list [1]>
#> 11  1.13      1     3     5 <named list [1]>
#> 12  0.148     2     3     5 <named list [1]>
#> 13  2.37      3     3     5 <named list [1]>
#> 14  0.784     4     3     5 <named list [1]>
#> 15  0.814     5     3     5 <named list [1]>

ssd_generate_data(ssdtools::ssd_rlnorm, nrow = 5, nsims = 3)
#> # A tibble: 15 × 5
#>      Conc   row   sim  nrow args            
#>     <dbl> <int> <int> <dbl> <list>          
#>  1  1.02      1     1     5 <named list [1]>
#>  2  1.03      2     1     5 <named list [1]>
#>  3  1.73      3     1     5 <named list [1]>
#>  4  0.103     4     1     5 <named list [1]>
#>  5 14.6       5     1     5 <named list [1]>
#>  6  0.697     1     2     5 <named list [1]>
#>  7  1.24      2     2     5 <named list [1]>
#>  8  2.93      3     2     5 <named list [1]>
#>  9  0.514     4     2     5 <named list [1]>
#> 10  3.05      5     2     5 <named list [1]>
#> 11  0.782     1     3     5 <named list [1]>
#> 12  0.308     2     3     5 <named list [1]>
#> 13  0.377     3     3     5 <named list [1]>
#> 14  2.90      4     3     5 <named list [1]>
#> 15  1.14      5     3     5 <named list [1]>