Fit SSD Distributions to Simulated Data

Fit SSD Distributions to Simulated Data

Usage

ssd_fit_dists_sims(
  x,
  dists = ssdtools::ssd_dists_bcanz(),
  ...,
  seed = NULL,
  silent = TRUE,
  .progress = FALSE
)

Arguments

x

A data frame with sim and stream integer columns and a list column of the data frames to fit distributions to.

dists

A character vector of the distribution names.

...

Additional arguments passed to ssdtools::ssd_fit_dists().

seed

An integer of the starting seed or NULL.

silent

A flag indicating whether fits should fail silently.

.progress

Whether to show a purrr::progress bar.

Value

The x tibble with a list column fits of fistdist objects.